imported/w3c/web-platform-tests/fs/FileSystemBaseHandle-postMessage-MessagePort-windows.https.window.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=250499 rdar://104154652
Reviewed by Tim Nguyen.

In current implementation, message-target.html could post a "LOADED" message to itself. However, it does not know how to
handle "LOADED" message, and it would send a "ERROR" message to the source (itself) when it receives an unexpected
message (as shown in add_message_event_handlers() in message-target.js). Since it does not know how to handle "ERROR"
message too, it would keep posting this message to itself. For the flaky test, this means the child window created in a
previous subtest could become the entry global object for code in a subsequent subtest, leading to unexpected result.
For example, parsed url in window.open() can change when entry global object is different; see window open steps in
https://html.spec.whatwg.org/multipage/nav-history-apis.html#window-open-steps.

To fix this issue, make sure message-target.html does not post "LOADED" message to itself.

* LayoutTests/imported/w3c/web-platform-tests/fs/FileSystemBaseHandle-postMessage-MessagePort-windows.https.window-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/fs/FileSystemBaseHandle-postMessage-windows.https.window-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/fs/resources/message-target.html:

Canonical link: https://commits.webkit.org/258840@main