GamepadHapticActuator.playEffect() with a magnitude < 0.1 has no effect https://bugs.webkit.org/show_bug.cgi?id=250522
Reviewed by Youenn Fablet.

GamepadHapticActuator.playEffect() with a magnitude < 0.1 has no effect in
Safari but triggers vibration in Chrome. This is causing the gamepad to
vibrate a lot less with Safari than Chrome when playing games such as
Forza on XBox Cloud.

I noticed that passing a value < 0.1 as intensity to GameController doesn't
result in any vibration of the gamepad. To address this we now scale the
magnitude value to be in the [0.1; 1] range before passing it as intensity
to the GameController framework. The exception is the value 0 which is still
treated as 0 (no vibration).

This makes the experience on XBox Cloud a lot more comparable between Safari
and Chrome.

* Source/WebCore/platform/gamepad/cocoa/GameControllerHapticEffect.mm:
(WebCore::magnitudeToIntensity):
(WebCore::GameControllerHapticEffect::create):

Canonical link: https://commits.webkit.org/258874@main