[MQ4] Allow negative values in media queries https://bugs.webkit.org/show_bug.cgi?id=250647 rdar://104278059
Reviewed by Antoine Quint.

"Some media features with a 'range' type are said to be false in the negative range. This means that negative values are
valid and must be parsed, and that querying whether the media feature is equal to, less than, or less or equal than any
such negative value must evaluate to false. Querying whether the media feature is greater, or greater or equal, than
a negative value evaluates to true if the relationship is true."

- https://www.w3.org/TR/mediaqueries-4/#false-in-the-negative-range

In practice this applies to all range features for types other than 'ratio'.

* LayoutTests/TestExpectations:
* LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-w3c-1.html:
* LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-w3c-2.html:
* LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-w3c-3.html:
* LayoutTests/fast/dom/HTMLImageElement/sizes/image-sizes-w3c-4.html:
* LayoutTests/fast/media/mq-resolution-expected.txt:
* LayoutTests/fast/media/mq-resolution.html:
* LayoutTests/imported/w3c/web-platform-tests/css/mediaqueries/test_media_queries-expected.txt:
* Source/WebCore/css/query/GenericMediaQueryParser.cpp:
(WebCore::MQ::GenericMediaQueryParserBase::validateFeatureAgainstSchema):

Allow negative values during parsing. They already evaluate correctly.

Canonical link: https://commits.webkit.org/258938@main