imported/w3c/web-platform-tests/mimesniff/mime-types/charset-parameter.window.html  is flaky https://bugs.webkit.org/show_bug.cgi?id=250917 rdar://99409762
Reviewed by Tim Nguyen.

imported/w3c/web-platform-tests/mimesniff/mime-types/charset-parameter.window.html
is flaky on Cocoa ports due to loads intermittently failing with a
cfurlErrorNetworkConnectionLost error.

I shown the logs to the CFNetwork team and the error occurs when trying to reuse
connections. It seems the server doesn't support reusing connection and we reach
a retry limit in CFNetwork (rdar://100873042).

I believe the reason we're unable to reuse the connection is because the python
script uses `response.close_connection = True`. The recommendation from the
CFNetwork team is for the python script to also send the `Connection: close`
header so that CFNetwork doesn't attempt to reuse it.

I tried the suggested change and it seems to address the flakiness locally.

* LayoutTests/imported/w3c/web-platform-tests/mimesniff/mime-types/resources/mime-charset.py:
(main):

Canonical link: https://commits.webkit.org/259159@main